+++ interactive_engine/groot-module/src/main/java/com/alibaba/graphscope/groot/store/StoreService.java
@@ -326,7 +326,6 @@ public class StoreService implements MetricsAgent {
                     }
                     logger.info("ingest data [{}] complete", path);
                 });
-        logger.info("ingest all data complete");
     }
 
     private void ingestDataInternal(
